moving App.js and App.css to components
