+++ src/components/App.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import './App.css';
-import Display from './components/Display';
-import ButtonPanel from './components/ButtonPanel';
+import Display from './Display';
+import ButtonPanel from './ButtonPanel';
 
 function App() {
   return (
     <div className="App">
-        <Display />
-        <ButtonPanel />
+      <Display />
+      <ButtonPanel />
     </div>
   );
 }
